Navbar functionality updated for login and logout button
