+++ client/src/App.css
@@ -1,0 +1,3 @@
+.App {
+    margin-top: 70px;
+}
